Create Top10_countries
I used the Country column from the country table and the customer_id column from the customer table. I added COUNT to make sure the customer ids were counted for each country in my final result. I INNER JOINED all the primary keys that linked the customer table to the address table, the address table to the city table, and the city table to the country table because the Customer table and the country table aren’t linked. I made sure to group the count by country, and ordered the list by customer id numbers from highest to lowest. This allowed me to limit the data to the top 10 rows.